backend/pyproject.toml
[tool.poetry]
name = "connect"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
SQLAlchemy = "^1.4.37"
asyncpg = "^0.25.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.18.3"
aiosqlite = "^0.17.0"
httpx = "^0.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
